refactor(idea_service): extract select_idea (#163)
  before: 2080
  after:  1981 (-99)

321 tests pass.
